requirements.txt
Django==2.0.4
django-smart-selects==1.2.9
django-admin-list-filter-dropdown==1.0.3
django-rest-framework==0.1.0
django-ckeditor==4.5.3
django-rest-framework==0.1.0
django-rest-multiple-models==1.8.1